Create FIR Filter Using Basic Principle in MATLAB
You have an audio file which is contaminated by a 60 Hz hum. This is a common problem in audio applications where 60 Hz signals from power supply and other sources appear in audio recordings. Your job is to create a filter which will remove this unwanted signal but keep the maximum amount of the remaining desired audio signal.